posts.repository.js의 코드의 명확성을 위해 일부 코드를 수정하고 posts.service.js로 옮김

+++ services/posts.service.js
@@ -19,7 +19,7 @@ class PostService {
   };
   getPosts = async ({ userId }) => {
     try {
-      const posts = await this.postRepository.getPosts({ userId });
+      const posts = await this.postRepository.getUserPosts({ userId });
 
       if (!posts || posts.length == 0) {
         throw new Error("게시글이 존재하지 않습니다");
@@ -33,14 +33,21 @@ class PostService {
   };
   updatePost = async ({ userId, title, content, postId }) => {
     try {
-      const result = await this.postRepository.updatePost({
-        userId,
-        title,
-        content,
-        postId,
-      });
+      console.log("postId", postId);
+      console.log("userId", userId);
 
-      return result;
+      const postFind = await this.postRepository.findUniquePost(userId);
+
+      if (postFind.userId == userId) {
+        const message = await this.postRepository.updatePost({
+          title,
+          content,
+          postId,
+        });
+        return message;
+      } else {
+        throw new Error("작성자만 게시글을 수정할 수 있습니다");
+      }
     } catch (error) {
       console.log("[Service] 게시글 수정 실패");
       throw new Error(error.message || "서비스에서 에러 발생");
@@ -48,12 +55,19 @@ class PostService {
   };
   deletePost = async ({ postId, userId }) => {
     try {
-      const result = await this.postRepository.deletePost({
-        postId,
-        userId,
-      });
+      console.log("postId", postId);
+      console.log("userId", userId);
 
-      return result;
+      const postFind = await this.postRepository.findUniquePost(userId);
+
+      if (postFind.userId == userId) {
+        const message = await this.postRepository.deletePost({
+          postId,
+        });
+        return message;
+      } else {
+        throw new Error("작성자만 게시글을 삭제할 수 있습니다");
+      }
     } catch (error) {
       console.log("[Service] 게시글 삭제 실패");
       throw new Error(error.message || "서비스에서 에러 발생");
@@ -61,13 +75,13 @@ class PostService {
   };
   createPost = async ({ title, content, userId }) => {
     try {
-      const result = await this.postRepository.createPost({
+      const message = await this.postRepository.createPost({
         title,
         content,
         userId,
       });
 
-      return result;
+      return message;
     } catch (error) {
       console.log("[Service] 게시글 등록 실패");
       throw new Error(error.message || "서비스에서 에러 발생");
